fix: naive whole number tokens

+++ llm_context_builder/tokenizers/naive.py
@@ -3,4 +3,4 @@ from .base import Tokenizer
 
 class NaiveTokenizer(Tokenizer):
     def count_tokens(self, text):
-        return len(text) / 4
+        return int(len(text) / 4)
